Update to normalisation and determine_resonances
We now use the NHDS normalisation, but multiplied by om**2. This improves stability for solutions with very small om.
In addition, the resonance check in the relativistic case occurs now in (pperp,ppar) space, not in (Gamma,pparbar) space. This means the standard integration is used when the resonances are outside the momentum-space domain.