Update of VDFs and generate_distribution
A new fit type has been included: 6 - bi-Moyal distribution

The Moyal distribution can also be created in generate_distributions.f90 now

The handling of the kappa distribution in the fitting is changed so that a fit parameter for the perpendicular thermal speed exists (not only the perp_corr factor)

generate_distribution.f90 has a number of additional options (autoscaleS, maxPperpS, maxPparS). If autoscale=T, the code works as before. If autoscale=F, the code now overrides the maximum perpendicular and parallel momenta with the values in maxPperpS and maxPparS.

The input files are updated to reflect these changes.